Crate opus

High-level bindings for libopus.

Only brief descriptions are included here. For detailed information, consult the libopus documentation.

Modules

packet
Analyze raw Opus packets.

Structs

Decoder
An Opus decoder with associated state.

Encoder
An Opus encoder with associated state.

Error
An error generated by the Opus library.

Repacketizer
A repacketizer used to merge together or split apart multiple Opus packets.

RepacketizerState
An in-progress repacketization.

SoftClip
Soft-clipping to bring a float signal within the [-1,1] range.

Enums

Bandwidth
The available bandwidth level settings.

Channels
The available channel setings.

CodingMode
The possible coding modes for the codec.

ErrorCode
Possible error codes.

Functions

version
Get the libopus version string.

Type Definitions

Result
Opus error Result alias.