Module opus::packet

Analyze raw Opus packets.

Structs

Packet
A parsed Opus packet, retuned from parse.

Functions

get_bandwidth
Get the bandwidth of an Opus packet.

get_nb_channels
Get the number of channels from an Opus packet.

get_nb_frames
Get the number of frames in an Opus packet.

get_nb_samples
Get the number of samples of an Opus packet.

get_samples_per_frame
Get the number of samples per frame from an Opus packet.

pad
Pad a given Opus packet to a larger size.

parse
Parse an Opus packet into one or more frames.

unpad
Remove all padding from a given Opus packet and rewrite the TOC sequence to minimize space usage.